chore: remove fastify as dependency and add @fastify/cookie in advance
This effectively reverts ac461cc5, since it is decided that it is better
for fastify to be a peer dependency of @nestjs/platform-fastify to
simplify dependency management, despite a slight inconvenience of
not having fastify type completions.

This commit also adds @fastify/cookie as a dependency in advance, and it
will be used in the near future.